Disentangle tests from package test
